Mega Update
Added Support For:
Driving to a given coordinate (based on xy system of enclosure).
Driving onto the base correctly
Traversing the enclosure

Updated:
A new movement function go was created. Added support for different exit status which are returned when it breaks out of while loop. Exit status information can be found in the function.

New Features:
orientation of the robot is now included. Orientation is based on cardinal points (N,S,E,W) and is used to make sure the robot updates coordinate information properly.

Comments are in the code, text me if you want more of an explanation and I can even create a video if needed.